DL-12464 fixing it tests and warns and move itsettings config
