--- project/TestPhases.scala
@@ -1,6 +1,0 @@
-import sbt._
-
-private object TestPhases {
-  lazy val TemplateTest = config("tt") extend Test
-  lazy val TemplateItTest = config("tit") extend IntegrationTest
-}
